nest dev commands inside local
