+++ MANIFEST.in
@@ -1,5 +1,6 @@
 include VERSION.txt
 include hokusai/cli/*
+include hokusai/cli/local/*
 include hokusai/cli/remote/*
 include hokusai/commands/*
 include hokusai/lib/*
