+++ hokusai/cli/__init__.py
@@ -1,4 +1,5 @@
 from hokusai.cli.base import base
 from hokusai.cli.local import local
+from hokusai.cli.local.dev import dev
 from hokusai.cli.remote import remote
 from hokusai.cli.remote.env import env
