+++ hokusai/cli/local/__init__.py
@@ -50,56 +50,3 @@ def push(tag, force, overwrite, verbose):
   """Build and push an image to the project's remote repo tagged as the git SHA1 of HEAD"""
   set_verbosity(verbose)
   hokusai.push(tag, force, overwrite)
-
-
-@local.command(context_settings=CONTEXT_SETTINGS)
-@click.option('-b', '--build', type=click.BOOL, is_flag=True, help="Force rebuild the docker image before running")
-@click.option('-d', '--detach', type=click.BOOL, is_flag=True, help="Run containers in the background")
-@click.option('-v', '--verbose', type=click.BOOL, is_flag=True, help='Verbose output')
-def start(build, detach, verbose):
-  """Start the development stack defined in ./hokusai/development.yml"""
-  set_verbosity(verbose)
-  hokusai.dev_start(build, detach)
-
-
-@local.command(context_settings=CONTEXT_SETTINGS)
-@click.option('-v', '--verbose', type=click.BOOL, is_flag=True, help='Verbose output')
-def stop(verbose):
-  """Stop the development stack defined in ./hokusai/development.yml"""
-  set_verbosity(verbose)
-  hokusai.dev_stop()
-
-
-@local.command(context_settings=CONTEXT_SETTINGS)
-@click.option('-v', '--verbose', type=click.BOOL, is_flag=True, help='Verbose output')
-def status(verbose):
-  """Print the status of the development stack defined in ./hokusai/development.yml"""
-  set_verbosity(verbose)
-  hokusai.dev_status()
-
-
-@local.command(context_settings=CONTEXT_SETTINGS)
-@click.option('-v', '--verbose', type=click.BOOL, is_flag=True, help='Verbose output')
-@click.option('-f', '--follow', type=click.BOOL, is_flag=True, help="Follow output")
-@click.option('-t', '--tail', type=click.INT, help="Number of lines of recent logs to display")
-def logs(follow, tail, verbose):
-  """Print logs from the development stack defined in ./hokusai/development.yml"""
-  set_verbosity(verbose)
-  hokusai.dev_logs(follow, tail)
-
-
-@local.command(context_settings=CONTEXT_SETTINGS)
-@click.argument('command')
-@click.option('-v', '--verbose', type=click.BOOL, is_flag=True, help='Verbose output')
-def run(command, verbose):
-  """Run a command in the development stack's container with the name 'project-name' in hokusai/config.yml"""
-  set_verbosity(verbose)
-  hokusai.dev_run(command)
-
-
-@local.command(context_settings=CONTEXT_SETTINGS)
-@click.option('-v', '--verbose', type=click.BOOL, is_flag=True, help='Verbose output')
-def clean(verbose):
-  """Stop and remove all containers in the development stack defined in ./hokusai/development.yml"""
-  set_verbosity(verbose)
-  hokusai.dev_clean()
